fix build to always start fresh
Telegex.Hook.Server is only compiled if Plug is available. It seems
like, if the dep already got compiled once before Plug was added to
dependencies, it doesn't recompile and therefore Telegex.Hook.Server
won't be available. I don't want to deal with stuff like this and it
took me HOURS to figure this out so let's rather start fresh on every
build.